Update: made changes to doc and test cases
